RSR-2200 commit changes for version 2.2.9

+++ components/com_publications/models/curation.php
@@ -2024,6 +2024,7 @@ class Curation extends Obj
 		$second   = $this->getElements(2);
 		$gallery  = $this->getElements(3);
 		$elements = array_merge($prime, $second, $gallery);
+		Lang::load('com_publications', Component::path('com_publications') . '/site');
 
 		// Do we have items to package?
 		if (!$elements)
@@ -2087,28 +2088,27 @@ class Curation extends Obj
 	 * @param	boolean	$includeVersionNum
 	 * @return  mixed  False on error, string on success
 	 */
-	public function getBundleName($includeVersionNum = false)
+	public function getBundleName($symLinkName = false)
 	{
 		if (empty($this->_pub))
 		{
 			return false;
 		}
-		$doi = $this->_pub->version->get('doi');
-		if ($doi != '')
+		$bundleName = 'Publication' . '_' . $this->_pub->id;
+		if ($symLinkName)
 		{
-			$doi = str_replace('.', '_', $doi);
-			$doi = str_replace('/', '_', $doi);
-			$bundleName = $doi;
+			$bundleName .= '_' . $this->_pub->version->get('version_number');
 		}
 		else
 		{
-			$bundleName = Lang::txt('Publication') . '_' . $this->_pub->id;
-			if ($includeVersionNum)
+			$doi = $this->_pub->version->get('doi');
+			if ($doi != '')
 			{
-				$bundleName .= '_' . $this->_pub->version->get('version_number');
+				$doi = str_replace('.', '_', $doi);
+				$doi = str_replace('/', '_', $doi);
+				$bundleName = $doi;
 			}
 		}
-
 		return $bundleName . '.zip';
 	}
 
@@ -2158,15 +2158,20 @@ class Curation extends Obj
 	public function createSymLink()
 	{
 		$tarname = $this->getBundleName();
-		$tarpath = $this->_pub->path('base', true) . DS . $tarname;
+		$tarpath = $this->_pub->path('relative') . DS . $tarname;
 		$symLink = $this->_symLinkPath();
+		if ($symLink !== false)
+		{
+			chdir(dirname($symLink));
+		}
+
 		if (empty($this->_pub) || $symLink == false || !is_file($tarpath))
 		{
 			return false;
 		}
 		if (!is_file($symLink))
 		{
-			if (!symlink($tarpath, $symLink))
+			if (!link($tarpath, $symLink))
 			{
 				return false;
 			}
@@ -2186,7 +2191,11 @@ class Curation extends Obj
 		{
 			return false;
 		}
-		unlink($symLink);
+
+		if (is_file($symLink))
+		{
+			unlink($symLink);
+		}
 		return true;
 	}
